+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 ## unreleased
 *************
 - Added aggregation rule descriptions to the user guide.
+- Updated SPF healthstate to match the ICD.
+  - Now [`UNKNOWN`, `NORMAL`, `DEGRADED`, `FAILED`]
+  - DishManager `healthstate` is unchanged
 - Add interlockack command
 - Add new FannedOutTangoLongRunningCommand
   - Uses lrcQueue, lrcExecuting, and lrcFinished from SKA devices to update the status of commands.